backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
openai==1.51.0
PyMuPDF==1.24.11
python-docx==1.1.2
chromadb==0.5.15
reportlab==4.2.5
